Get rid of extraneous stuff.

+++ Install/Jenkins/jobs/cloneJobDefinition.sh
@@ -22,16 +22,15 @@ else
 	exit 1
 fi
 
-GIT_INSTALL_DIR="/usr/bin/git/bin"
-if [ ! -e "/usr/local/git/bin" ]; then
-	GIT_INSTALL_DIR="/usr/local/git/bin"
-fi
-
 # Clone the Job Definition
 if [ ! -e "${JENKINS_HOME}/jobs/${JOB_NAME}" ]; then
-    echo "${GIT_INSTALL_DIR}/git clone git://github.com/avendasora/${JOB_DEFINITION} ${JENKINS_HOME}/jobs/${JOB_NAME}"
-	(git clone git://github.com/avendasora/${JOB_DEFINITION} ${JENKINS_HOME}/jobs/${JOB_NAME})
-	curl -O ${JENKINS_URL}/reload
+    echo "git clone git://github.com/avendasora/${JOB_DEFINITION} ${JENKINS_HOME}jobs/${JOB_NAME}"
+	git clone git://github.com/avendasora/${JOB_DEFINITION} ${JENKINS_HOME}jobs/${JOB_NAME}
+	if [ -e "${JENKINS_HOME}jobs/${JOB_NAME}" ]; then
+		curl -O ${JENKINS_URL}/reload
+	else
+		exit 1
+	fi
 else
 	echo "A job already exists with the name ${JOB_NAME}. You must delete the existing job, or chose a different name."
 	exit 1
